data_request_app/requirements.txt
pandas
requests
regex
mysql-connector-python
sqlalchemy
PyMySQL